Switch to using global layout listener
If a showcase view is created and shown in onCreate(), the
showcase view's width/height may still be zero, when onPreDraw()
is called. This causes the reveal radius to be 0, so the circular
reveal doesn't show.

Switching to the global layout listener seems to fix this. It is
the same technique used for the focus view mView if has not been
measured.